pyproject.toml
[project]
name = "trade-platform"
version = "0.1.0"
description = "Production-ready FastAPI service for a trading platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.0",
  "asyncpg>=0.30.0",
  "cryptography>=45.0.0",
  "fastapi>=0.116.0",
  "gunicorn>=23.0.0",
  "greenlet>=3.1.0",
  "ccxt>=4.5.0",
  "numpy>=2.2.0",
  "pandas>=2.3.0",
  "pandas-ta>=0.3.14b0",
  "pydantic-settings>=2.10.0",
  "python-dotenv>=1.1.0",
  "httpx>=0.28.0",
  "redis>=6.0.0",
  "sqlalchemy>=2.0.41",
  "taskiq>=0.11.0",
  "taskiq-redis>=1.0.0",
  "uvicorn[standard]>=0.35.0",
  "python-jose[cryptography]>=3.5.0",
  "pwdlib[argon2]>=0.3.0",
  "email-validator>=2.3.0",
]

[dependency-groups]
dev = [
  "aiosqlite>=0.22.1",
  "mypy>=1.17.0",
  "pandas-stubs>=2.3.0.250703",
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
  "ruff>=0.12.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
show_error_codes = true
pretty = true
plugins = []

[[tool.mypy.overrides]]
module = [
  "alembic.*",
  "taskiq.*",
  "taskiq_redis.*",
  "ccxt.*",
  "pandas_ta.*",
  "jose",
  "jose.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.alembic]
script_location = "migrations"

[project.scripts]
dev = "app.scripts:dev"
lint = "app.scripts:lint"
format = "app.scripts:format_code"
typecheck = "app.scripts:typecheck"
test = "app.scripts:test"
